gravel: rename 'orch' controller to 'ceph'
