+++ src/aquarium.py
@@ -24,10 +24,10 @@ from fastapi.staticfiles import StaticFiles
 
 from gravel.api import auth, devices, local, nodes, orch, status, users
 from gravel.cephadm.cephadm import Cephadm
+from gravel.controllers.ceph.ceph import Ceph, Mgr, Mon
 from gravel.controllers.gstate import GlobalState, setup_logging
 from gravel.controllers.inventory.inventory import Inventory
 from gravel.controllers.nodes.mgr import NodeMgr
-from gravel.controllers.orch.ceph import Ceph, Mgr, Mon
 from gravel.controllers.resources.devices import Devices
 from gravel.controllers.resources.status import Status
 from gravel.controllers.resources.storage import Storage
