+++ src/gravel/controllers/resources/storage.py
@@ -18,9 +18,9 @@ from fastapi.logger import logger as fastapi_logger
 from pydantic.fields import Field
 from pydantic.main import BaseModel
 
+from gravel.controllers.ceph.ceph import Mon
 from gravel.controllers.gstate import Ticker
 from gravel.controllers.nodes.mgr import NodeMgr
-from gravel.controllers.orch.ceph import Mon
 
 logger: Logger = fastapi_logger
