+++ src/gravel/tests/unit/controllers/ceph/data/device_ls_not_available.json
