+++ src/gravel/tests/unit/controllers/ceph/data/mon_osdmap_raw.json
